Bug 11132: lib/log_proc: ensure the same signedness when comparing numbers
This patch gets rid of a compilation warning.

+++ lib/log_proc/log_filters_yaml.c
@@ -60,7 +60,7 @@ get_level_mask(const char *level_str)
 
         for (i = 0; i < levels_num; i++)
         {
-            if (len == strlen(levels[i].name) &&
+            if ((size_t)len == strlen(levels[i].name) &&
                 strncmp(ptr, levels[i].name, len) == 0)
             {
                 val |= levels[i].mask;
